Fixed Data Issues.
