+++ PasswordPolicySystem.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PolicyLibrary", "PolicyLibr
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PolicyLibrary.Tests", "PolicyLibrary.Tests\PolicyLibrary.Tests.csproj", "{BBE08315-BE68-4DC0-A994-EC445391DEE1}"
 EndProject
+Project("{00D1A9C2-B5F0-4AF3-8072-F6C62B433612}") = "LocalDb", "LocalDb\LocalDb.sqlproj", "{14689575-75C6-4D1A-92C2-D216BD6B24D5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,12 @@ Global
 		{BBE08315-BE68-4DC0-A994-EC445391DEE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BBE08315-BE68-4DC0-A994-EC445391DEE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BBE08315-BE68-4DC0-A994-EC445391DEE1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14689575-75C6-4D1A-92C2-D216BD6B24D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14689575-75C6-4D1A-92C2-D216BD6B24D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14689575-75C6-4D1A-92C2-D216BD6B24D5}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{14689575-75C6-4D1A-92C2-D216BD6B24D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14689575-75C6-4D1A-92C2-D216BD6B24D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14689575-75C6-4D1A-92C2-D216BD6B24D5}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
