+++ WebLoginDemo/Shared/MainLayout.razor
@@ -3,7 +3,8 @@
 <div class="page">
     <div class="main">
         <div class="top-row px-4">
-            
+            <a href="/">Go Home</a>
+            <a href="/all">View User Data</a>
         </div>
 
         <div class="content px-4">
