+++ WebLoginDemo/appsettings.json
@@ -7,6 +7,9 @@
     }
   },
   "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "PolicyDb": "Data Source=(localdb)\\MSSQLLocalDB;Integrated Security=True;Persist Security Info=False;Pooling=False;MultipleActiveResultSets=False;Connect Timeout=60;Encrypt=False;TrustServerCertificate=False"
+  },
   "PolicySettings": {
     "Policy": "PartOne"
   }
